+++ timetracker/cfg/doc_local.py
@@ -9,14 +9,11 @@ in a version supported by cygwin, conda, and venv.
 __copyright__ = 'Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.'
 __author__ = "DV Klopfenstein, PhD"
 
+import os.path as op
 from os.path import dirname
-from os.path import normpath
-from os.path import basename
 from collections import namedtuple
 from glob import glob
-from timetracker.cfg.utils import get_username
-from timetracker.cfg.utils import get_abspath
-from timetracker.cfg.utils import replace_envvar
+from timetracker.cfg import utils
 from timetracker.cfg.tomutils import read_config
 from timetracker.cfg.docutils import get_ntvalue
 from timetracker.starttime import Starttime
@@ -47,7 +44,7 @@ class DocProj:
         assert type(doc).__name__ != 'RdCfg'
         assert filename is not None
         self.filename = filename
-        self.dircfg  = normpath(dirname(filename))
+        self.dircfg  = op.normpath(dirname(filename))
         self.project, self.csv_filename, self.global_config_filename, self.errors = \
             self._init_cfg_values(doc)
         self.dircsv = dirname(self.csv_filename) if self.csv_filename else None
@@ -56,24 +53,24 @@ class DocProj:
         """Get the absolute pathname for doc['csv']['filename']"""
         dirproj = dirname(self.dircfg)
         if self.dircsv is not None and dirproj is not None:
-            return get_abspath(self.dircsv, dirproj)
+            return utils.get_abspath(self.dircsv, dirproj)
         return None
 
     def get_filename_csv(self, username=None, dirhome=None):
         """Get the csv filename by reading the cfg csv pattern and filling in"""
         assert username is None or '/' not in username
-        username = get_username(username)
+        username = utils.get_username(username)
         return self._get_csvfilename_proj_user(username, dirhome)
 
     def get_filenames_csv(self, dirhome):
         """Get the csv filename by reading the cfg csv pattern and filling in"""
         if (fcsvpat := self._get_csvfilename_proj(dirhome)) is not None:
-            return glob(replace_envvar(fcsvpat, '*'))
+            return glob(utils.replace_envvar(fcsvpat, '*'))
         return None
 
     def get_csv_username(self, fcsv):
         """Using the csv pattern in the project config, determine the username"""
-        bname = basename(self.csv_filename)
+        bname = op.basename(self.csv_filename)
         pc0 = bname.find('$')
         if pc0 == -1:
             return None
@@ -105,7 +102,7 @@ class DocProj:
     def get_startobj(self, username):
         """Get a Starttime object"""
         if self.project:
-            return Starttime(self.dircfg, self.project, get_username(username))
+            return Starttime(self.dircfg, self.project, utils.get_username(username))
         return None
 
     ##-------------------------------------------------------------
@@ -113,13 +110,13 @@ class DocProj:
         """Read a config file and load it into a TOML document"""
         fcsvpat = self._get_csvfilename_proj(dirhome)
         if fcsvpat:
-            return replace_envvar(fcsvpat, username) if '$' in fcsvpat else fcsvpat
+            return utils.replace_envvar(fcsvpat, username) if '$' in fcsvpat else fcsvpat
         return None
 
     def _get_csvfilename_proj(self, dirhome):
         """Read a config file and load it into a TOML document"""
         if self.csv_filename and self.project:
-            fpat = get_abspath(self.csv_filename, dirname(self.dircfg), dirhome)
+            fpat = utils.get_abspath(self.csv_filename, dirname(self.dircfg), dirhome)
             return fpat.replace('PROJECT', self.project)
         return None