+++ timetracker/cmd/cancel.py
@@ -3,9 +3,8 @@
 __copyright__ = 'Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.'
 __author__ = "DV Klopfenstein, PhD"
 
-from timetracker.msgs import str_cancelled1
-from timetracker.msgs import str_not_running
-from timetracker.cmd.common import prt_elapsed
+from timetracker import msgs
+from timetracker.cmd import common
 from timetracker.cfg.cfg_local import CfgProj
 
 
@@ -18,10 +17,10 @@ def cli_run_cancel(fnamecfg, args):
 def run_cancel(cfgproj, name=None):
     """Cancel a timer if it is started"""
     if (startobj := cfgproj.get_starttime_obj(name)) and startobj.started():
-        prt_elapsed(startobj, f'{str_cancelled1()}; was')
+        common.prt_elapsed(startobj, f'{msgs.str_cancelled1()}; was')
         startobj.rm_starttime()
         return startobj.filename
-    print(str_not_running())
+    print(msgs.str_not_running())
     return None