+++ timetracker/docx.py
@@ -3,25 +3,9 @@
 __copyright__ = 'Copyright (C) 2025-present, DV Klopfenstein, PhD. All rights reserved.'
 __author__ = "DV Klopfenstein, PhD"
 
-##from os import remove
-#from os.path import exists
-##from os.path import basename
-##from os.path import join
-##from os.path import abspath
-##from os.path import dirname
-##from os.path import normpath
 ##https://python-docx.readthedocs.io/en/latest/
 from docx import Document
 from docx.shared import Inches
-#from docx.shared import Inches
-#from datetime import timedelta
-#from datetime import datetime
-#from logging import debug
-#
-#from timetracker.utils import orange
-#from timetracker.consts import DIRTRK
-##from timetracker.cfg.utils import get_username
-
 
 class WordDoc:
     """Generate a Microsoft Word document containing a table of data"""